+++ src/library/DIPS.Mobile.UI/Effects/Touch/iOS/TouchPlatformEffect.cs
@@ -56,12 +56,14 @@ public partial class TouchPlatformEffect
     {
         if (m_touchMode is Touch.TouchMode.Tap or Touch.TouchMode.Both)
         {
-            Control.RemoveGestureRecognizer(m_tapGestureRecognizer!);
+            if(Control.GestureRecognizers != null)
+                Control.RemoveGestureRecognizer(m_tapGestureRecognizer!);
         }
 
         if (m_touchMode is Touch.TouchMode.LongPress or Touch.TouchMode.Both)
         {
-            Control.RemoveGestureRecognizer(m_longPressGestureRecognizer!);
+            if(Control.GestureRecognizers != null)
+                Control.RemoveGestureRecognizer(m_longPressGestureRecognizer!);
         }
     }